Update and rename game.java to Game.java
